+++ yash-env/src/system/concurrency/delegates.rs
@@ -468,23 +468,6 @@ where
     }
 }
 
-// Concurrent<S> cannot implement Fork because the return type of the fork
-// method does not match with that of the inner system S. Instead, Concurrent<S>
-// provides an inherent method `new_child_process` that returns a
-// `ChildProcessStarter<S>` as returned by the inner method.
-// impl<S> Fork for Concurrent<S>
-// where
-//     S: Fork,
-// {
-//     #[inline]
-//     fn new_child_process(&self) -> Result<ChildProcessStarter<Self>>
-//     where
-//         Self: Sized,
-//     {
-//         todo!()
-//     }
-// }
-
 impl<S> Concurrent<S>
 where
     S: Fork,